The users credits are shown on the nav bar. Did a clean up of contentAccess.mjs
